+++ src/js/widgets.js
@@ -21,4 +21,5 @@ module.exports = [
     require( '../widget/analog-scale/analog-scalepicker' ),
     require( '../widget/big-image/image-viewer' ),
     require( '../widget/comment/commentwidget' ),
+    require( '../widget/image-map/image-map' )
 ];